final commit

+++ test/integration/InteractionsTest.t.sol
@@ -1,12 +1,13 @@
 //SPDX-License-Identifier:MIT
 pragma solidity ^0.8.18;
+
 import {Test, console} from "forge-std/Test.sol";
 import {FundMe} from "../../src/FundMe.sol";
 import {DeployFundMe} from "../../script/DeployFundMe.s.sol";
 import {FundFundMe, WithdrawFundMe} from "../../script/Interactions.s.sol";
 
 contract InteractionsTest is Test {
-    uint number = 1;
+    uint256 number = 1;
     FundMe fundMe;
     address User = makeAddr("User"); //this is only used in foundry(solidity) to make fake addresses
     uint256 constant Gas_Price = 1;
